refactor(table): add useEditableRow
